+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [v1.0.0] - 2026-05-28
+## [v1.1.0] - 2026-05-28
+
+### Added
+- JSON configuration file support (`~/.sshell/config`)
+- Configuration file options: default_user, default_port, default_auth, default_alive, verbose
+
+### Fixed
+- N/A
+
+### Changed
+- N/A
 
 > 📖 [中文说明](docs/CHANGELOG_zh-CN.md)